feat: refactor SegmentNumber widget for consistent segment number rendering
