+++ lib/shared/utils/helper_functions.dart
@@ -52,11 +52,18 @@ double? getFontSize(String language) {
   }
 }
 
-/// Converts ⤵ to <br> for content passed to an HTML renderer.
-String normalizeSegmentHtml(String? raw) => raw?.replaceAll('⤵', '<br>') ?? '';
+/// Soft line-break marker embedded in segment content by the backend.
+/// Replaced with a renderer-appropriate break by [normalizeSegmentHtml] /
+/// [normalizeSegmentText].
+const String kSegmentSoftBreak = '⤵';
 
-/// Converts ⤵ to \n for content displayed in a plain Text widget.
-String normalizeSegmentText(String? raw) => raw?.replaceAll('⤵', '\n') ?? '';
+/// Converts [kSegmentSoftBreak] to <br> for content passed to an HTML renderer.
+String normalizeSegmentHtml(String? raw) =>
+    raw?.replaceAll(kSegmentSoftBreak, '<br>') ?? '';
+
+/// Converts [kSegmentSoftBreak] to \n for content displayed in a plain Text widget.
+String normalizeSegmentText(String? raw) =>
+    raw?.replaceAll(kSegmentSoftBreak, '\n') ?? '';
 
 /// Calculates the share position origin for share_plus ShareParams.
 ///
